+++ stencil-workspace/src/components/modus-toast/modus-toast.tsx
@@ -23,7 +23,14 @@ export class ModusToast {
   @Prop() showIcon = true;
 
   /** (optional) The toasts' type. */
-  @Prop() type: 'danger' | 'dark' | 'default' | 'primary' | 'secondary' | 'success' | 'tertiary' | 'warning' = 'default';
+  @Prop() type:
+    | 'danger'
+    | 'dark'
+    | 'default'
+    | 'primary'
+    | 'secondary'
+    | 'success'
+    | 'warning' = 'default';
 
   /** An event that fires when the toast is dismissed */
   @Event() dismissClick: EventEmitter;
@@ -61,7 +68,9 @@ export class ModusToast {
           <slot />
         </span>
         <span class={'close'}>
-          {this.dismissible && <IconClose size={'18'} onClick={() => this.dismissClick.emit()} />}
+          {this.dismissible && (
+            <IconClose size={'18'} onClick={() => this.dismissClick.emit()} />
+          )}
         </span>
       </div>
     );